Fix an off-by-one bug in the alignment extraction when only a single sequence is submitted
In the case of a single sequence, the bogus sequence wasn't being
submitted and so results didn't match up in parse_html.

Always add the bogus sequence even if there's more than one submitted
sequence.  It's the second time I wrote buggy code because I tested two
sequences but didn't test one sequence and the new code didn't account
for the bogus sequence.  It's easier just to always account for it one
way or another.  It's also a good reason for unit tests, but those don't
exist (yet).